NOT fix + unsigned support + fix tests

+++ c_compiler/driver/float_funct.c
@@ -1,6 +1,6 @@
 
 float f(){
-    int z = 4;
+    float z = 4;
     float x =6.75;  
     x = ++z ;
     return z+x;
